# src/main/webapp/api-docs/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "0.0.1-SNAPSHOT",
    "title" : "API's do Libros"
  },
  "host" : "localhost:8080/lpoow_jboss",
  "basePath" : "/api",
  "tags" : [ {
    "name" : "bibliotecas"
  }, {
    "name" : "clientes"
  }, {
    "name" : "livros"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/bibliotecas" : {
      "get" : {
        "tags" : [ "bibliotecas" ],
        "summary" : "Recuperar as bibliotecas",
        "description" : "",
        "operationId" : "listar",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/BibliotecaDto"
              }
            }
          }
        }
      }
    },
    "/bibliotecas/{idBiblioteca}" : {
      "get" : {
        "tags" : [ "bibliotecas" ],
        "summary" : "Recuperar uma biblioteca",
        "description" : "",
        "operationId" : "recuperar",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idBiblioteca",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/BibliotecaDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "bibliotecas" ],
        "summary" : "Remover uma biblioteca",
        "description" : "",
        "operationId" : "remover",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idBiblioteca",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/bibliotecas/{idBiblioteca}/atualizar/{nomeBiblioteca}" : {
      "put" : {
        "tags" : [ "bibliotecas" ],
        "summary" : "Atualizar uma biblioteca",
        "description" : "",
        "operationId" : "atualizar",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idBiblioteca",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "nomeBiblioteca",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/BibliotecaDto"
            }
          }
        }
      }
    },
    "/bibliotecas/{nomeBiblioteca}" : {
      "post" : {
        "tags" : [ "bibliotecas" ],
        "summary" : "Adicionar uma biblioteca",
        "description" : "",
        "operationId" : "adicionar",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "nomeBiblioteca",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/clientes" : {
      "get" : {
        "tags" : [ "clientes" ],
        "summary" : "Recuperar os clientes",
        "description" : "",
        "operationId" : "listar",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ClienteDto"
              }
            }
          }
        }
      }
    },
    "/clientes/{cpfCliente}/{nomeCliente}" : {
      "post" : {
        "tags" : [ "clientes" ],
        "summary" : "Adicionar um cliente",
        "description" : "",
        "operationId" : "adicionar",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "cpfCliente",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "nomeCliente",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/clientes/{idCliente}" : {
      "get" : {
        "tags" : [ "clientes" ],
        "summary" : "Recuperar um cliente",
        "description" : "",
        "operationId" : "recuperar",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idCliente",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ClienteDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "clientes" ],
        "summary" : "Remover um cliente",
        "description" : "",
        "operationId" : "remover",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idCliente",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/clientes/{idCliente}/atualizar/{cpfCliente}/{nomeCliente}" : {
      "put" : {
        "tags" : [ "clientes" ],
        "summary" : "Atualizar um cliente",
        "description" : "",
        "operationId" : "atualizar",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idCliente",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "cpfCliente",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "nomeCliente",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ClienteDto"
            }
          }
        }
      }
    },
    "/livros" : {
      "get" : {
        "tags" : [ "livros" ],
        "summary" : "Recuperar os livros",
        "description" : "",
        "operationId" : "listar",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/LivroDto"
              }
            }
          }
        }
      }
    },
    "/livros/{idLivro}" : {
      "get" : {
        "tags" : [ "livros" ],
        "summary" : "Recuperar um livro",
        "description" : "",
        "operationId" : "recuperar",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idLivro",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/LivroDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "livros" ],
        "summary" : "Remover um livro",
        "description" : "",
        "operationId" : "remover",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idLivro",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/livros/{idLivro}/atualizar/{isbnLivro}/{nomeLivro}" : {
      "put" : {
        "tags" : [ "livros" ],
        "summary" : "Atualizar um livro",
        "description" : "",
        "operationId" : "atualizar",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idLivro",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "isbnLivro",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "nomeLivro",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/LivroDto"
            }
          }
        }
      }
    },
    "/livros/{isbnLivro}/{nomeLivro}" : {
      "post" : {
        "tags" : [ "livros" ],
        "summary" : "Adicionar um livro",
        "description" : "",
        "operationId" : "adicionar",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "isbnLivro",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "nomeLivro",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "definitions" : {
    "BibliotecaDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nome" : {
          "type" : "string"
        },
        "livros" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Livro"
          }
        }
      }
    },
    "ClienteDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "cpf" : {
          "type" : "string"
        },
        "nome" : {
          "type" : "string"
        },
        "livros" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Livro"
          }
        }
      }
    },
    "Livro" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "isbn" : {
          "type" : "string"
        },
        "nome" : {
          "type" : "string"
        }
      }
    },
    "LivroDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "isbn" : {
          "type" : "string"
        },
        "nome" : {
          "type" : "string"
        }
      }
    }
  }
}